# docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "realestate",
        "title": "realestate",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
        "/v1/properties/list": {
            "post": {
                "description": "List properties by ids.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "properties"
                ],
                "summary": "List properties by ids.",
                "parameters": [
                    {
                        "description": "property list body",
                        "name": "Search",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/propertyListRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/DefaultResponseContainer"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/propertyResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/v1/saved-homes/{id}": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Adds property to user saved homes",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "savedhomes"
                ],
                "summary": "Add saved home",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/propertyResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Delete user saved home.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "savedhomes"
                ],
                "summary": "Delete user saved home.",
                "responses": {
                    "204": {
                        "description": "No Content",
                        "schema": {
                            "$ref": "#/definitions/EmptyResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/v1/saved-searches/{id}": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Update UserSearchFilters by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "searchfilters"
                ],
                "summary": "Update UserSearchFilters",
                "parameters": [
                    {
                        "description": "search filters data",
                        "name": "searchFilters",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/searchFiltersRequest"
                        }
                    },
                    {
                        "type": "integer",
                        "description": "UserSearchFilters filters ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content",
                        "schema": {
                            "$ref": "#/definitions/EmptyResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Delete user search filter.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "searchfilters"
                ],
                "summary": "Delete user search filter.",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Search filters ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content",
                        "schema": {
                            "$ref": "#/definitions/EmptyResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "DefaultResponseContainer": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "data": {},
                "totalCount": {
                    "type": "integer"
                }
            }
        },
        "EmptyResponse": {
            "type": "object"
        },
        "ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "address": {
            "type": "object",
            "properties": {
                "city": {
                    "type": "string"
                },
                "country": {
                    "type": "string"
                },
                "houseNumber": {
                    "type": "string"
                },
                "neighborhood": {
                    "type": "string"
                },
                "state": {
                    "type": "string"
                },
                "street": {
                    "type": "string"
                },
                "zipCode": {
                    "type": "string"
                }
            }
        },
...
        "polygon": {
            "type": "object",
            "properties": {
                "botLat": {
                    "type": "string"
                },
                "botLong": {
                    "type": "string"
                },
                "topLat": {
                    "type": "string"
                },
                "topLong": {
                    "type": "string"
                }
            }
        },
...
        "propertyListRequest": {
            "type": "object",
            "required": [
                "ids"
            ],
            "properties": {
                "ids": {
                    "type": "array",
                    "maxItems": 10,
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
...
        "propertyLocationHome": {
            "type": "object",
            "properties": {
                "lat": {
                    "type": "number"
                },
                "long": {
                    "type": "number"
                }
            }
        },
...
        "propertyResponse": {
            "type": "object",
            "properties": {
                "actualDays": {
                    "type": "integer"
                },
                "address": {
                    "$ref": "#/definitions/address"
                },
                "appliance": {
                    "type": "boolean"
                },
                "bathroom": {
                    "type": "integer"
                },
                "bedroom": {
                    "type": "integer"
                },
                "brokerName": {
                    "type": "string"
                },
                "condition": {
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "floor": {
                    "type": "integer"
                },
                "fullAddress": {
                    "type": "string"
                },
                "has3DTour": {
                    "type": "boolean"
                },
                "hasAC": {
                    "type": "boolean"
                },
                "hasGarage": {
                    "type": "boolean"
                },
                "hasImages": {
                    "type": "boolean"
                },
                "hasVideo": {
                    "type": "boolean"
                },
                "heating": {
                    "type": "integer"
                },
                "homeSize": {
                    "type": "number"
                },
                "homeType": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "images": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "isActiveStatus": {
                    "type": "boolean"
                },
                "isSavedHome": {
                    "type": "boolean"
                },
                "location": {
                    "$ref": "#/definitions/propertyLocationHome"
                },
                "lotSize": {
                    "type": "number"
                },
                "petsAllowed": {
                    "type": "boolean"
                },
                "price": {
                    "type": "number"
                },
                "priceCurrency": {
                    "type": "string"
                },
                "propertyType": {
                    "type": "integer"
                },
                "totalFloors": {
                    "type": "integer"
                },
                "totalParking": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                },
                "yearBuild": {
                    "type": "integer"
                }
            }
        },
...
        "savedSearchFilters": {
            "type": "object",
            "properties": {
                "bathroom": {
                    "type": "integer"
                },
                "bathroomExact": {
                    "type": "boolean"
                },
                "bedroom": {
                    "type": "integer"
                },
                "bedroomExact": {
                    "type": "boolean"
                },
                "condition": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "hasAC": {
                    "type": "boolean"
                },
                "homeSizeFrom": {
                    "type": "number"
                },
                "homeSizeTo": {
                    "type": "number"
                },
                "homeType": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "lotSizeFrom": {
                    "type": "number"
                },
                "lotSizeTo": {
                    "type": "number"
                },
                "mustHaveGarage": {
                    "type": "boolean"
                },
                "parkingNumber": {
                    "type": "integer"
                },
                "priceFrom": {
                    "type": "number"
                },
                "priceTo": {
                    "type": "number"
                },
                "propertyType": {
                    "type": "integer"
                },
                "yearBuildFrom": {
                    "type": "integer"
                },
                "yearBuildTo": {
                    "type": "integer"
                }
            }
        },
...
        "searchFiltersRequest": {
            "type": "object",
            "properties": {
                "coordinatesRect": {
                    "$ref": "#/definitions/polygon"
                },
                "emailFrequency": {
                    "type": "integer",
                    "maximum": 10,
                    "minimum": 1
                },
                "filterValues": {
                    "$ref": "#/definitions/savedSearchFilters"
                },
                "name": {
                    "type": "string",
                    "maxLength": 300
                },
                "sort": {
                    "type": "integer",
                    "maximum": 10,
                    "minimum": 1
                },
                "subscribed": {
                    "type": "boolean"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}